fixup! Add more negative checks for integers passed to OPENSSL_malloc().

+++ crypto/bio/bf_buff.c
@@ -291,7 +291,7 @@ static long buffer_ctrl(BIO *b, int cmd, long num, void *ptr)
         if (num > ctx->ibuf_size) {
             if (num < 0)
                 return 0;
-            p1 = OPENSSL_malloc((int)num);
+            p1 = OPENSSL_malloc((size_t)num);
             if (p1 == NULL)
                 goto malloc_error;
             OPENSSL_free(ctx->ibuf);
@@ -322,12 +322,12 @@ static long buffer_ctrl(BIO *b, int cmd, long num, void *ptr)
         if ((ibs > DEFAULT_BUFFER_SIZE) && (ibs != ctx->ibuf_size)) {
             if (num < 0)
                 return 0;
-            p1 = OPENSSL_malloc((int)num);
+            p1 = OPENSSL_malloc((size_t)num);
             if (p1 == NULL)
                 goto malloc_error;
         }
         if ((obs > DEFAULT_BUFFER_SIZE) && (obs != ctx->obuf_size)) {
-            p2 = OPENSSL_malloc((int)num);
+            p2 = OPENSSL_malloc((size_t)num);
             if (p2 == NULL) {
                 if (p1 != ctx->ibuf)
                     OPENSSL_free(p1);
